// frontend/src/sections/items/detail/ancillaryItems.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type ItemAncillaryItemsVariables = {
  storeId: string;
  itemId: string;
};

export type ItemAncillaryItemsResult = {
  items: ({
  __typename: "ItemConnector";
} & {
  nodes: Array<{
  ancillaryItems: Array<{
  id: string;
  itemQuantity: number;
  ancillaryQuantity: number;
  ancillaryItem: {
  id: string;
  code: string;
  name: string;
} | null;
}>;
}>;
});
};

export const ItemAncillaryItems = {
  query: "query itemAncillaryItems($storeId: String!, $itemId: String!) {\n  items(storeId: $storeId, filter: {id: {equalTo: $itemId}}) {\n    ... on ItemConnector {\n      __typename\n      nodes {\n        ancillaryItems {\n          id\n          itemQuantity\n          ancillaryQuantity\n          ancillaryItem {\n            id\n            code\n            name\n          }\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<ItemAncillaryItemsResult, ItemAncillaryItemsVariables>;